requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.2.1
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloud-sql-python-connector==1.12.1
cramjam==2.9.1
cryptography==44.0.0
dataclasses-json==0.6.7
deptry==0.23.0
distro==1.9.0
et_xmlfile==2.0.0
fastparquet==2024.11.0
firebase-admin==6.6.0
Flask==2.2.2
Flask-Bcrypt==0.7.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.4.3
Flask-SQLAlchemy==3.0.2
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
jellyfish==1.1.3
Jinja2==3.1.5
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==2.1.5
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.1.0
mypy_extensions==1.1.0
#mysqlclient==2.2.7
narwhals==1.25.0
numpy==1.26.2
openai==1.79.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.1.4
plotly==6.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
recordclass==0.22.1
recordlinkage==0.16
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
requirements-parser==0.11.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.1.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tqdm==4.67.1
types-setuptools==75.8.0.20250110
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==2.2.2
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.15.2
zipp==3.20.2
zstandard==0.23.0
